fix(options): use float64 type in Options.Threshold

+++ options.go
@@ -198,7 +198,7 @@ type Options struct {
 	Compression    int
 	Zoom           int
 	Crop           bool
-	SmartCrop      bool // Deprecated
+	SmartCrop      bool // Deprecated, use: bimg.Options.Gravity = bimg.GravitySmart
 	Enlarge        bool
 	Embed          bool
 	Flip           bool
@@ -220,6 +220,6 @@ type Options struct {
 	Interpretation Interpretation
 	GaussianBlur   GaussianBlur
 	Sharpen        Sharpen
-	Threshold      float32
+	Threshold      float64
 	OutputICC      string
 }
